Refactor user roles management by moving USER_ROLES to a new constants file and updating its usage in survey.ts and persona.ts. Enhance USER_ROLES with icons for UI display.
