+++ apps/web/app/(landing)/welcome/survey.ts
@@ -18,89 +18,34 @@ import {
   UserIcon,
   CircleHelpIcon,
 } from "lucide-react";
+import { USER_ROLES as USER_ROLES_DATA } from "@/utils/constants/user-roles";
 
-export const USER_ROLES = [
-  {
-    value: "Founder",
-    description: "Building a startup or running my own company",
-    icon: RocketIcon,
-  },
-  {
-    value: "Executive",
-    description: "C-level, VP, or Director managing teams",
-    icon: BriefcaseIcon,
-  },
-  {
-    value: "Small Business Owner",
-    description: "Running a local business or solo venture",
-    icon: StoreIcon,
-  },
-  {
-    value: "Software Engineer",
-    description: "Writing code and building software",
-    icon: CodeIcon,
-  },
-  {
-    value: "Assistant",
-    description: "Managing communications and calendars for others",
-    icon: CalendarDaysIcon,
-  },
-  {
-    value: "Investor",
-    description: "VC, angel investor, or fund manager",
-    icon: TrendingUpIcon,
-  },
-  {
-    value: "Sales",
-    description: "Closing deals and managing client relationships",
-    icon: PhoneIcon,
-  },
-  {
-    value: "Marketing",
-    description: "Growing brands and driving campaigns",
-    icon: MegaphoneIcon,
-  },
-  {
-    value: "Customer Support",
-    description: "Helping customers and resolving issues",
-    icon: HeadphonesIcon,
-  },
-  {
-    value: "Realtor",
-    description: "Buying, selling, and managing properties",
-    icon: HomeIcon,
-  },
-  {
-    value: "Content Creator",
-    description: "YouTuber, blogger, or social media influencer",
-    icon: VideoIcon,
-  },
-  {
-    value: "Consultant",
-    description: "Advising businesses and solving problems",
-    icon: UsersIcon,
-  },
-  {
-    value: "E-commerce",
-    description: "Running an online store or marketplace",
-    icon: ShoppingCartIcon,
-  },
-  {
-    value: "Student",
-    description: "Studying at school or university",
-    icon: GraduationCapIcon,
-  },
-  {
-    value: "Individual",
-    description: "Managing my personal email",
-    icon: UserIcon,
-  },
-  {
-    value: "Other",
-    description: "My role isn't listed here",
-    icon: CircleHelpIcon,
-  },
-] as const;
+// Add icons to the plain user roles data for UI display
+export const USER_ROLES = USER_ROLES_DATA.map((role) => {
+  const iconMap: Record<string, any> = {
+    Founder: RocketIcon,
+    Executive: BriefcaseIcon,
+    "Small Business Owner": StoreIcon,
+    "Software Engineer": CodeIcon,
+    Assistant: CalendarDaysIcon,
+    Investor: TrendingUpIcon,
+    Sales: PhoneIcon,
+    Marketing: MegaphoneIcon,
+    "Customer Support": HeadphonesIcon,
+    Realtor: HomeIcon,
+    "Content Creator": VideoIcon,
+    Consultant: UsersIcon,
+    "E-commerce": ShoppingCartIcon,
+    Student: GraduationCapIcon,
+    Individual: UserIcon,
+    Other: CircleHelpIcon,
+  };
+
+  return {
+    ...role,
+    icon: iconMap[role.value],
+  };
+});
 
 export const survey = {
   questions: [
